requirements_all.txt
attrs==19.3.0
numpy==1.18.0
cython==0.29.14
rhasspy-asr==0.1.0
flake8==3.7.9
pylint==2.4.4
